29/8

+++ resources/views/admin/order/list.blade.php
@@ -17,7 +17,7 @@
                 @endif
                 <table class="table table-bordered">
                     <thead>
-                        <tr>
+                        <tr style="text-align: center">
                             <th>STT</th>
                             <th>Tên người dùng</th>
                             <th>Số điện thoại</th>
@@ -25,7 +25,7 @@
                             <th>Tổng giá đơn hàng</th>
                             <th>Ngày tạo đơn hàng</th>
                             <th>Trạng thái</th>
-                            <th>Xử lý đơn</th>
+                            <th colspan="2">Xử lý đơn</th>
                             <th>Thao tác</th>
                         </tr>
                     </thead>
@@ -38,23 +38,41 @@
                                 <td>{{ $item->address }}</td>
                                 <td>{{ number_format($item->total_price) }}đ</td>
                                 <td>{{ $item->created_at }}</td>
-                                <td>{{ $item->status }}</td>
+                                @if ($item->status == 0)
+                                    <td>Đang chờ duyệt</td>
+                                @elseif($item->status == 1)
+                                    <td>Đang giao hàng</td>
+                                @else
+                                    <td>Hoàn thành</td>
+                                @endif
                                 <td
                                     style="display:flex; width: 100% ; justify-content: space-between;border-top: none;border-left: none;border-right: none">
-                                    <form action="{{ route('deleteProduct', $item->id) }}" method="POST">
-                                        @csrf
-                                        <button type="submit" class="btn btn-primary">Xác nhận thanh toán</button>
-                                    </form>
-                                    <form action="{{ route('deleteProduct', $item->id) }}" method="POST">
-                                        @csrf
-                                        <button type="submit" class="btn btn-danger">Hủy đơn</button>
-                                    </form>
-                                    <form action="{{ route('delete_item', $item->id) }}" method="POST">
-                                        @csrf
-                                        @method('DELETE')
-                                        <button onclick="return confirm('bạn có chắc muốn xóa ?')" type="submit"
-                                            class="btn btn-danger">Xóa</button>
-                                    </form>
+                                    @if ($item->status == 1)
+                                        <form action="{{ route('update_status', $item->id) }}" method="POST">
+                                            @csrf
+                                            @method('PUT')
+                                            <button type="submit" class="btn btn-primary">Xác nhật thanh toán</button>
+                                        </form>
+                                    @elseif($item->status == 2)
+                                        <button disabled type="submit" class="btn btn-default">Đơn hàng đã hoàn
+                                            thành</button>
+                                    @else
+                                        <form action="{{ route('update_status', $item->id) }}" method="POST">
+                                            @csrf
+                                            @method('PUT')
+                                            <button type="submit" class="btn btn-success">Duyệt đơn hàng</button>
+                                        </form>
+                                    @endif
+                                </td>
+                                <td>
+                                    @if ($item->status == 0 || $item->status == 1)
+                                        <form action="{{ route('delete_item', $item->id) }}" method="POST">
+                                            @csrf
+                                            @method('DELETE')
+                                            <button onclick="return confirm('bạn có chắc muốn xóa ?')" type="submit"
+                                                class="btn btn-danger">Hủy đơn</button>
+                                        </form>
+                                    @endif
                                 </td>
                                 <td style="">
                                     <a onclick="" href="{{ route('order_Detail', $item->id) }}"
